COMMERCE-6073 Cast dsl query return to avoid class cast exception when using GetterUtil

+++ modules/apps/commerce/commerce-account-service/src/main/java/com/liferay/commerce/account/service/impl/CommerceAccountLocalServiceImpl.java
@@ -464,7 +464,7 @@ public class CommerceAccountLocalServiceImpl
 		throws PortalException {
 
 		return GetterUtil.getInteger(
-			accountEntryLocalService.dslQuery(
+			(Long)accountEntryLocalService.dslQuery(
 				_getGroupByStep(
 					DSLQueryFactoryUtil.countDistinct(
 						AccountEntryTable.INSTANCE.accountEntryId),
